MacOS: copy the plugins from neubot-0.4.15.6.pkg
I don't have the 10.6 SDK in this Mac. I probably can install a
copy of it quite easily, but I'm a bit lazy today.

Therefore I have fetched a copy of the 0.4.15.6 package, and I
have copied the plugins from it.

Also remove the code to build the privacy plugin from makepkg.py,
because now we the files we need are already in the template.